FINAL_QI_2025/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
2captcha-python==1.2.2
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
altgraph==0.17.4
annotated-types==0.6.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
asyncio==3.4.3
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
bidict==0.23.1
blinker==1.7.0
bs4==0.0.1
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.7
cssselect==1.2.0
cycler==0.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.3.1
debugpy==1.8.0
decorator==5.1.1
discord-webhook==1.3.1
distlib==0.3.7
distro==1.8.0
Django==5.0.3
dnspython==2.4.2
email-validator==2.1.0.post1
env==0.1.0
et-xmlfile==1.1.0
executing==2.0.1
fake-useragent==1.4.0
fastjsonschema==2.19.0
feedparser==6.0.11
filelock==3.13.1
firebase-admin==6.3.0
flake8==7.0.0
Flask==3.0.0
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.44.0
frozendict==2.3.8
frozenlist==1.4.1
fsspec==2024.3.0
get-chrome-driver==1.3.19
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.1
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
install==1.3.5
ipykernel==6.26.0
ipython==8.17.2
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
locket==1.0.0
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mccabe==0.7.0
msgpack==1.0.7
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
nbformat==5.9.2
nest-asyncio==1.5.8
npm==0.1.1
numpy==1.26.1
openai==0.27.0
openpyxl==3.1.2
optional-django==0.1.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
parse==1.20.0
parso==0.8.3
partd==1.4.1
passlib==1.7.4
patsy==0.5.6
peewee==3.17.0
pefile==2023.2.7
pexpect==4.8.0
Pillow==10.1.0
pip-tools==7.3.0
pipenv==2023.10.24
platformdirs==3.11.0
plotly==5.18.0
pluggy==1.3.0
pretty-html-table==0.9.16
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-postgresql==1.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyee==8.2.2
pyflakes==3.2.0
Pygments==2.16.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyJWT==2.8.0
PyMySQL==1.1.0
pypandoc==1.12
pyparsing==3.1.1
pyppeteer==1.0.2
pyproject_hooks==1.0.0
pyquery==2.0.0
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-engineio==4.9.0
python-socketio==5.11.1
python-utils==3.8.1
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.32.0
requests==2.31.0
requests-html==0.10.0
retrying==1.3.4
rpds-py==0.13.2
rsa==4.9
schedule==1.2.1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
selenium==4.16.0
setuptools==68.2.2
sgmllib3k==1.0.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.1
svgwrite==1.4.3
tenacity==8.2.3
threadpoolctl==3.2.0
tomlkit==0.12.4
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
Tree==0.2.4
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.24.6
w3lib==2.1.2
waitress==2.1.2
wcwidth==0.2.9
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.1
wheel==0.41.3
wsproto==1.2.0
WTForms==3.1.1
XlsxWriter==3.2.0
yahoo-fin==0.8.9.1
yarl==1.9.4
yfinance==0.2.31
zipp==3.17.0